removing files that were moved to MEF
