+++ src/index.js
@@ -1,8 +1,7 @@
-const { icareApp, zipErrors } = require('./cliUtils');
+const { icareApp } = require('./app');
 const { generateNewMessageBundle } = require('./icareFhirMessaging');
 
 module.exports = {
   generateNewMessageBundle,
   icareApp,
-  zipErrors,
 };
